fix: move fonts css variables to layout
